prevent out of memory
if the tcp transport receives a malicious packet (or a packet not
originating from a fastbin server/client) the size field can contain any
kind of int. For very high int values this will lead to OOM errors.

Adds a max size of (default) 10MB to prevent a DOS attack angle.